feat: adds flags to specify version and directory ike is downloaded to

+++ scripts/get.sh
@@ -12,6 +12,17 @@ die () {
 TAR_FILE="/tmp/ike.tar.gz"
 RELEASES_URL="https://github.com/maistra/istio-workspace/releases"
 
+show_help() {
+  echo "get - downloads ike binary matching your operating system"
+  echo " "
+  echo "./get.sh [options]"
+  echo " "
+  echo "Options:"
+  echo "-h, --help          shows brief help"
+  echo "-v, --version       defines version specific version of the binary to download (defaults to latest)"
+  echo "-d, --dir           target directory to which the binary is downloaded (defaults to random tmp dir in /tmp suffixed with ike-version)"
+}
+
 last_version() {
   curl -sL -o /dev/null -w %{url_effective} "$RELEASES_URL/latest" |
     rev |
@@ -20,19 +31,66 @@ last_version() {
 }
 
 download() {
+  version=$1
+  if [[ ${version} == "" ]]; then
+    echo >&2 "Undefined version (pass using -v|--version). Please use semantic version. Read more about it here: https://semver.org/ \n\n"
+    show_help
+    exit 1
+  fi
+
+  url="$RELEASES_URL/download/$version/ike_${version:1}_$(uname -s)_$(uname -m).tar.gz"
+
   rm -f "$TAR_FILE"
-  curl -s -L -o "$TAR_FILE" \
-    "$RELEASES_URL/download/$VERSION/ike_${VERSION:1}_$(uname -s)_$(uname -m).tar.gz"
+  curl -fsLo "$TAR_FILE" "$url" || die "Unable to download $url"
 }
 
-test -z "$VERSION" && VERSION="$(last_version)"
-test -z "$VERSION" && {
-  die "Unable to get ike version. You can still try to download manually from $RELEASES_URL"
+while test $# -gt 0; do
+  case "$1" in
+    -h|--help)
+            show_help
+            exit 0
+            ;;
+    -v)
+            shift
+            if test $# -gt 0; then
+              version=$1
+            else
+              die "Please provide a version name"
+            fi
+            shift
+            ;;
+    --version*)
+            version=`echo $1 | sed -e 's/^[^=]*=//g'`
+            shift
+            ;;
+    -d)
+            shift
+            if test $# -gt 0; then
+              dir=$1
+            else
+              die "Please provide a version name"
+            fi
+            shift
+            ;;
+    --dir*)
+            dir=`echo $1 | sed -e 's/^[^=]*=//g'`
+            shift
+            ;;
+    *)
+            die "Unknown param $1"
+            break
+            ;;
+  esac
+done
+
+test -z "$version" && version="$(last_version)"
+test -z "$version" && {
+  die "Unable to get ike version. You can still try to download it manually from $RELEASES_URL."
 }
-test -z "$TMPDIR" && TMPDIR="$(mktemp -d --suffix=-ike-${VERSION})"
+test -z "$dir" &&  dir="$(mktemp -d --suffix=-ike-${version})"
 
-download
-tar -xf "$TAR_FILE" -C "$TMPDIR"
+download ${version}
+tar -C "$dir" -xzf "$TAR_FILE" ike
 
-echo "Downloaded ike binary to $TMPDIR"
-echo -e "You can add it to your path by typing following in your terminal:\n$ export PATH=\"\$PATH:$TMPDIR\""
+echo "Downloaded ike binary ($version) to $dir"
+echo "Make sure it's on your \$PATH."
